fix after code review, small fix data init
